Update WayFinder.java
The update to the most recently revised code is as follows: The latest errors have been resolved, and additional code has been added to display error messages on the screen when the user makes mistakes. Moreover, there has been a fundamental change in the code. Additionally, a “Previous” array has been integrated into the code.